create the google authenticator class
just a small refactoring

the totp class shoudl be the base class and should not perform any special
treatment for inputs (unless the RFC specifically tells it to, like in the
case of the moving factor being a byte aray of size 8).

so we create a different class that will layer in these special behaviors
on top of the base class.